Untangle more dependencies
This allows us to make `clx_render_benchmark` and
`text_render_integration_test` standalone.
