+++ Source/controls/touch/renderers.cpp
@@ -7,6 +7,7 @@
 #include "engine/events.hpp"
 #include "engine/render/clx_render.hpp"
 #include "engine/render/primitive_render.hpp"
+#include "game_mode.hpp"
 #include "init.hpp"
 #include "inv.h"
 #include "levels/gendung.h"
