+++ Source/diablo.h
@@ -66,7 +66,7 @@ enum class MouseActionType : uint8_t {
 extern uint32_t DungeonSeeds[NUMLEVELS];
 extern DVL_API_FOR_TEST std::optional<uint32_t> LevelSeeds[NUMLEVELS];
 extern Point MousePosition;
-extern DVL_API_FOR_TEST bool gbRunGame;
+
 extern bool gbRunGameResult;
 extern bool ReturnToMainMenu;
 extern bool gbProcessPlayers;
