+++ Source/engine/load_cl2.hpp
@@ -12,7 +12,6 @@
 #include "engine/clx_sprite.hpp"
 #include "engine/load_file.hpp"
 #include "mpq/mpq_common.hpp"
-#include "utils/cl2_to_clx.hpp"
 #include "utils/endian_read.hpp"
 #include "utils/endian_write.hpp"
 #include "utils/pointer_value_union.hpp"
@@ -23,6 +22,8 @@
 #ifdef UNPACKED_MPQS
 #define DEVILUTIONX_CL2_EXT ".clx"
 #else
+#include "utils/cl2_to_clx.hpp"
+
 #define DEVILUTIONX_CL2_EXT ".cl2"
 #endif
