+++ Source/engine/render/clx_render.hpp
@@ -6,8 +6,6 @@
 #pragma once
 
 #include <cstdint>
-
-#include <array>
 #include <utility>
 
 #ifdef DEBUG_CLX
@@ -18,7 +16,6 @@
 #include "engine/point.hpp"
 #include "engine/render/light_render.hpp"
 #include "engine/surface.hpp"
-#include "lighting.h"
 
 namespace devilution {
 
@@ -96,39 +93,9 @@ void ClxDrawBlended(const Surface &out, Point position, ClxSprite clx);
  * @param out Output buffer
  * @param position Target buffer coordinate
  * @param clx CLX frame
- */
-inline void ClxDrawLight(const Surface &out, Point position, ClxSprite clx, int lightTableIndex)
-{
-	if (lightTableIndex != 0) {
-		ClxDrawTRN(out, position, clx, LightTables[lightTableIndex].data());
-	} else {
-		ClxDraw(out, position, clx);
-	}
-}
-
-/**
- * @brief Blit CL2 sprite, and apply lighting, to the given buffer at the given coordinates
- * @param out Output buffer
- * @param position Target buffer coordinate
- * @param clx CLX frame
  * @param lightmap Per-pixel light buffer
  */
 void ClxDrawWithLightmap(const Surface &out, Point position, ClxSprite clx, const Lightmap &lightmap);
-
-/**
- * @brief Blit CL2 sprite, and apply lighting and transparency blending, to the given buffer at the given coordinates
- * @param out Output buffer
- * @param position Target buffer coordinate
- * @param clx CLX frame
- */
-inline void ClxDrawLightBlended(const Surface &out, Point position, ClxSprite clx, int lightTableIndex)
-{
-	if (lightTableIndex != 0) {
-		ClxDrawBlendedTRN(out, position, clx, LightTables[lightTableIndex].data());
-	} else {
-		ClxDrawBlended(out, position, clx);
-	}
-}
 
 /**
  * @brief Blit CL2 sprite, and apply lighting and transparency blending, to the given buffer at the given coordinates
