+++ Source/engine/render/text_render.cpp
@@ -10,6 +10,7 @@
 #include <cstddef>
 #include <cstdint>
 #include <optional>
+#include <string_view>
 #include <utility>
 #include <variant>
 
@@ -17,23 +18,22 @@
 #include <fmt/core.h>
 
 #include "DiabloUI/ui_flags.hpp"
+#include "engine/clx_sprite.hpp"
 #include "engine/load_cel.hpp"
 #include "engine/load_clx.hpp"
 #include "engine/load_file.hpp"
 #include "engine/load_pcx.hpp"
-#include "engine/palette.h"
 #include "engine/point.hpp"
+#include "engine/rectangle.hpp"
 #include "engine/render/clx_render.hpp"
 #include "engine/render/primitive_render.hpp"
 #include "engine/surface.hpp"
 #include "engine/ticks.hpp"
-#include "options.h"
+#include "game_mode.hpp"
 #include "utils/algorithm/container.hpp"
-#include "utils/display.h"
 #include "utils/is_of.hpp"
 #include "utils/language.h"
 #include "utils/log.hpp"
-#include "utils/sdl_compat.h"
 #include "utils/utf8.hpp"
 
 namespace devilution {
@@ -490,20 +490,6 @@ uint32_t DoDrawString(const Surface &out, std::string_view text, Rectangle rect,
 	maybeDrawCursor();
 	return static_cast<uint32_t>(remaining.data() - text.data());
 }
-
-void OptionLanguageCodeChanged()
-{
-	UnloadFonts();
-	LanguageInitialize();
-	LoadLanguageArchive();
-	effects_cleanup_sfx();
-	if (gbRunGame)
-		sound_init();
-	else
-		ui_sound_init();
-}
-
-const auto OptionChangeHandlerResolution = (GetOptions().Language.code.SetValueChangedCallback(OptionLanguageCodeChanged), true);
 
 } // namespace
