+++ Source/engine/render/text_render.hpp
@@ -5,6 +5,7 @@
  */
 #pragma once
 
+#include <cstddef>
 #include <cstdint>
 #include <optional>
 #include <string>
@@ -13,11 +14,10 @@
 #include <variant>
 #include <vector>
 
-#include <SDL.h>
-
 #include "DiabloUI/ui_flags.hpp"
 #include "engine/clx_sprite.hpp"
 #include "engine/palette.h"
+#include "engine/point.hpp"
 #include "engine/rectangle.hpp"
 #include "engine/surface.hpp"
 #include "utils/enum_traits.h"
