+++ Source/game_mode.hpp
@@ -4,6 +4,8 @@
 
 namespace devilution {
 
+/* Are we in-game? If false, we're in the main menu. */
+extern DVL_API_FOR_TEST bool gbRunGame;
 /** Indicate if we only have access to demo data */
 extern DVL_API_FOR_TEST bool gbIsSpawn;
 /** Indicate if we have loaded the Hellfire expansion data */
