+++ Source/qol/monhealthbar.cpp
@@ -15,6 +15,7 @@
 #include "engine/load_clx.hpp"
 #include "engine/render/clx_render.hpp"
 #include "engine/render/primitive_render.hpp"
+#include "game_mode.hpp"
 #include "options.h"
 #include "utils/language.h"
 #include "utils/str_cat.hpp"
