+++ Source/utils/pcx_to_clx.cpp
@@ -6,6 +6,7 @@
 
 #include <array>
 #include <memory>
+#include <optional>
 #include <vector>
 
 #include <SDL_endian.h>
