+++ Source/utils/pcx_to_clx.hpp
@@ -3,6 +3,8 @@
 #include <cstdint>
 #include <optional>
 
+#include <SDL.h>
+
 #include "engine/assets.hpp"
 #include "engine/clx_sprite.hpp"
